Pin indexes to enum, relocated&renamed callback functions and files
Updated callback function for sensor reading interrupt to be more descriptive, renamed and relocated callback function starting pin reads to align with callback status. Added macros for 3.3v max and max ADC value. Created sensor index enum. Updated gitignore to remove settings changes from diff.
