+++ fsae-vehicle-fw/src/peripherals/adc.cpp
@@ -10,20 +10,9 @@ void ADC1_init() {
     adc->adc1->setResolution(12); // set bits of resolution
     adc->adc1->setConversionSpeed(ADC_CONVERSION_SPEED::LOW_SPEED); // change the conversion speed
     adc->adc1->setSamplingSpeed(ADC_SAMPLING_SPEED::LOW_SPEED); // change the sampling speed
-    adc->adc1->enableInterrupts(callbacks);
+    adc->adc1->enableInterrupts(ADCReadingCompleteCallback);
     Serial.println("Done initializing ADCs");
 }
 
 uint32_t begin;
 
-void ADC1_start() {
-    uint8_t currentPin = pins[adcIndex];
-    Serial.print( "Starting read for ");
-    Serial.println(currentPin);
-    Serial.print("time since prev start of read");
-    uint32_t currentTime = micros();
-    Serial.println(currentTime - begin);
-    begin = micros();
-    adc->adc1->startSingleRead(currentPin); // in callbacks.h
-
-}
