+++ fsae-vehicle-fw/src/peripherals/peripherals.cpp
@@ -1,6 +1,6 @@
 // Anteater Electric Racing, 2025
 
-#include "timer.h"
+#include "peripherals/timer.h"
 #include "peripherals/adc.h"
 #include "peripherals/gpio.h"
 #include "peripherals/peripherals.h"
@@ -10,6 +10,5 @@ void Peripherals_Init() {
     ADC1_init();
     Serial.print("initialized ADCs");
     // init the rest of peripherals
-    // ADC1_start();
     startTimer();
 }
