+++ fsae-vehicle-fw/src/peripherals/timer.cpp
@@ -1,0 +1,6 @@
+#include "peripherals/timer.h"
+#include "callbacks.h"
+
+void startTimer(){
+    int result = readPinTimer.begin(StartADCPinReadings, 1000);
+}
